+++ app/src/main/java/susu/stepanvolkov/shop/activity/CartActivity.kt
@@ -1,12 +1,11 @@
-package susu.stepanvolkov.shop.activities
+package susu.stepanvolkov.shop.activity
+
 
 import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
+import androidx.appcompat.app.AppCompatActivity
 import kotlinx.android.synthetic.main.activity_cart.*
-import kotlinx.android.synthetic.main.activity_cart.view.*
-import kotlinx.android.synthetic.main.activity_cart.view.checkoutOrderBtn
 import kotlinx.android.synthetic.main.toolbar_layout.view.*
 import susu.stepanvolkov.shop.R
 
@@ -19,8 +18,8 @@ class CartActivity : AppCompatActivity() {
         toolbar.headerText.text = getString(R.string.cart_header)
         toolbar.headerBackBtn.setOnClickListener{ finish() }
         toolbar.shoppingCartBtn.visibility = View.GONE
+
         checkoutOrderBtn.setOnClickListener{
-            // TODO: Переход к CheckoutActivity
             val intent = Intent(this, CheckoutActivity::class.java)
             startActivity(intent)
         }
