+++ app/src/main/java/susu/stepanvolkov/shop/activity/CatalogActivity.kt
@@ -1,4 +1,4 @@
-package susu.stepanvolkov.shop.activities
+package susu.stepanvolkov.shop.activity
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
