+++ app/src/main/res/layout/activity_cart.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:background="@color/main_background"
     android:layout_height="match_parent"
-    tools:context=".activities.CartActivity">
+    tools:context=".activity.CartActivity">
 
     <include
         android:id="@+id/toolbar"
@@ -36,6 +36,7 @@
                 android:textColor="@color/black"
                 android:text="@string/good_list_header"
                 android:textSize="18sp"
+                android:textStyle="bold"
                 android:paddingHorizontal="18dp"
                 android:paddingTop="36dp"
                 app:layout_constraintTop_toTopOf="parent"
@@ -69,7 +70,7 @@
                     android:id="@+id/orderHeader"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:layout_marginStart="32dp"
+                    android:layout_marginStart="16dp"
                     android:layout_marginTop="32dp"
                     android:text="@string/order_header"
                     android:textColor="@color/black"
@@ -82,7 +83,7 @@
                     android:id="@+id/fullPriceHeader"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:layout_marginStart="32dp"
+                    android:layout_marginStart="16dp"
                     android:layout_marginTop="16dp"
                     android:text="@string/full_price"
                     android:textColor="@color/black"
@@ -106,7 +107,7 @@
                     android:id="@+id/discountHeader"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:layout_marginStart="32dp"
+                    android:layout_marginStart="16dp"
                     android:layout_marginTop="16dp"
                     android:text="@string/discount"
                     android:textColor="@color/black"
@@ -130,7 +131,7 @@
                     android:id="@+id/priceWithDiscountHeader"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:layout_marginStart="32dp"
+                    android:layout_marginStart="16dp"
                     android:layout_marginTop="16dp"
                     android:text="@string/total_price"
                     android:textColor="@color/black"
