+++ app/src/main/res/layout/activity_catalog.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.CatalogActivity">
+    tools:context=".activity.CatalogActivity">
 
     <include
         android:id="@+id/toolbar"
