+++ app/src/main/res/layout/activity_checkout.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.CheckoutActivity">
+    tools:context=".activity.CheckoutActivity">
 
     <include
         android:id="@+id/toolbar"
