+++ app/src/main/res/layout/activity_product_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/main_background"
-    tools:context=".activities.ProductDetailActivity">
+    tools:context=".activity.ProductDetailActivity">
 
     <include
         android:id="@+id/toolbar"
